+++ apps/api-harmonization/src/blocks/service-list/service-list.request.ts
@@ -1,0 +1,10 @@
+import { CMS } from '@o2s/framework/modules';
+
+export class GetServiceListBlockQuery implements Omit<CMS.Request.GetCmsEntryParams, 'locale'> {
+    id!: string;
+    offset?: number;
+    limit?: number;
+    type?: string;
+    category?: string;
+    status?: string;
+}