+++ apps/frontend/src/components/Chart/ChartTooltip/ChartTooltip.tsx
@@ -1,16 +1,14 @@
 import { TooltipProps } from 'recharts';
 import { NameType, ValueType } from 'recharts/types/component/DefaultTooltipContent';
 
+import { Models } from '@o2s/framework/modules';
+
 import { Typography } from '@o2s/ui/components/typography';
 
-import { useGlobalContext } from '@/providers/GlobalProvider';
-
-import { Currency } from '@/hooks/usePriceService';
+import { Price } from '@/components/Price/Price';
 
 export const ChartTooltip = (props: TooltipProps<ValueType, NameType>) => {
     const { active, payload } = props;
-
-    const { priceService } = useGlobalContext();
 
     if (!active || !payload?.length) {
         return null;
@@ -29,7 +27,9 @@ export const ChartTooltip = (props: TooltipProps<ValueType, NameType>) => {
                                 <Typography variant="small">{`${bar?.name} :`}</Typography>
                             </div>
                             <Typography variant="small" className="text-right">
-                                {priceService.formatPrice({ value: Number(bar.value) }, bar.unit as Currency).format}
+                                <Price
+                                    price={{ value: Number(bar.value), currency: bar.unit as Models.Price.Currency }}
+                                />
                             </Typography>
                         </div>
                     ))