+++ packages/framework/src/modules/resources/resources.model.ts
@@ -1,30 +1,12 @@
 import { Pagination } from '@/utils/models';
 
-export enum ProductType {
-    PHYSICAL,
-    VIRTUAL,
-}
+export type ProductType = 'PHYSICAL' | 'VIRTUAL';
 
-export enum AssetStatus {
-    ACTIVE,
-    INACTIVE,
-    RETIRED,
-}
+export type AssetStatus = 'ACTIVE' | 'INACTIVE' | 'RETIRED';
 
-export enum ContractStatus {
-    ACTIVE,
-    EXPIRED,
-    INACTIVE,
-}
+export type ContractStatus = 'ACTIVE' | 'EXPIRED' | 'INACTIVE';
 
-export class Product {
-    id!: string;
-    name!: string;
-    description!: string;
-    url!: string;
-    type!: ProductType;
-    category!: string;
-}
+export type PaymentPeriod = 'ONE_TIME' | 'MONTHLY' | 'YEARLY';
 
 export class Contract {
     id!: string;
@@ -32,11 +14,12 @@ export class Contract {
     status!: ContractStatus;
     startDate!: string;
     endDate!: string;
+    paymentPeriod?: PaymentPeriod;
 }
 
 export class Resource {
     id!: string;
-    product!: Product;
+    productId!: string;
     billingAccountId!: string;
 }
 
@@ -55,3 +38,5 @@ export class Asset extends Resource {
 }
 
 export type Resources = Pagination.Paginated<Resource>;
+export type Services = Pagination.Paginated<Service>;
+export type Assets = Pagination.Paginated<Asset>;