+++ packages/integrations/mocked/src/modules/resources/resources.service.ts
@@ -3,20 +3,28 @@ import { Observable, of } from 'rxjs';
 
 import { Resources } from '@o2s/framework/modules';
 
-import { mapResource, mapResources } from './resources.mapper';
+import { mapAsset, mapAssets, mapService, mapServices } from './resources.mapper';
 import { responseDelay } from '@/utils/delay';
 
 @Injectable()
 export class ResourcesService implements Resources.Service {
-    getResourceList(query: Resources.Request.GetResourceListQuery): Observable<Resources.Model.Resources> {
-        return of(mapResources(query)).pipe(responseDelay());
-    }
-
-    getResource(params: Resources.Request.GetResourceParams): Observable<Resources.Model.Resource> {
-        return of(mapResource(params.id)).pipe(responseDelay());
-    }
-
     purchaseOrActivateResource(_params: Resources.Request.GetResourceParams): Observable<void> {
         throw new Error('Method not implemented');
+    }
+
+    getServiceList(query: Resources.Request.GetServiceListQuery): Observable<Resources.Model.Services> {
+        return of(mapServices(query)).pipe(responseDelay());
+    }
+
+    getService(params: Resources.Request.GetServiceParams): Observable<Resources.Model.Service> {
+        return of(mapService(params.id)).pipe(responseDelay());
+    }
+
+    getAssetList(query: Resources.Request.GetAssetListQuery): Observable<Resources.Model.Assets> {
+        return of(mapAssets(query)).pipe(responseDelay());
+    }
+
+    getAsset(params: Resources.Request.GetAssetParams): Observable<Resources.Model.Asset> {
+        return of(mapAsset(params.id)).pipe(responseDelay());
     }
 }