+++ packages/integrations/strapi-cms/src/modules/cms/mappers/cms.page.mapper.ts
@@ -26,6 +26,21 @@ export const mapPage = (data: PageFragment): CMS.Model.Page.Page => {
         hasOwnTitle: data.hasOwnTitle,
         parent: {
             slug: data.parent?.slug ?? '',
+            seo: {
+                title: data.parent?.SEO!.title ?? '',
+            },
+            parent: {
+                slug: data.parent?.parent?.slug ?? '',
+                seo: {
+                    title: data.parent?.parent?.SEO!.title ?? '',
+                },
+                parent: {
+                    slug: data.parent?.parent?.parent?.slug ?? '',
+                    seo: {
+                        title: data.parent?.parent?.parent?.SEO!.title ?? '',
+                    },
+                },
+            },
         },
     };
 };
@@ -52,6 +67,21 @@ export const mapAlternativePages = (data: PageFragment): CMS.Model.Page.Page => 
         hasOwnTitle: data.hasOwnTitle,
         parent: {
             slug: data.parent?.slug ?? '',
+            seo: {
+                title: data.parent?.SEO!.title ?? '',
+            },
+            parent: {
+                slug: data.parent?.parent?.slug ?? '',
+                seo: {
+                    title: data.parent?.parent?.SEO!.title ?? '',
+                },
+                parent: {
+                    slug: data.parent?.parent?.parent?.slug ?? '',
+                    seo: {
+                        title: data.parent?.parent?.parent?.SEO!.title ?? '',
+                    },
+                },
+            },
         },
     };
 };
@@ -119,5 +149,9 @@ const mapComponent = (component: ComponentFragment) => {
             return 'PaymentsHistoryBlock';
         case 'ComponentComponentsUserAccount':
             return 'UserAccountBlock';
+        case 'ComponentComponentsServiceList':
+            return 'ServiceListBlock';
+        case 'ComponentComponentsServiceDetails':
+            return 'ServiceDetailsBlock';
     }
 };