+++ packages/ui/src/components/link.tsx
@@ -10,6 +10,8 @@ const linkVariants = cva(
         variants: {
             variant: {
                 default: 'text-primary underline-offset-4 hover:underline',
+                primaryButton:
+                    'h-10 px-4 py-2 no-underline hover:no-underline bg-primary text-primary-foreground hover:bg-primary/90',
             },
         },
         defaultVariants: {