+++ apps/frontend/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @o2s/frontend
 
+## 0.18.0
+
+### Minor Changes
+
+- 98b2b61: implemented breadcrumbs
+- 77f9dc4: UI theme update
+
+### Patch Changes
+
+- Updated dependencies [77f9dc4]
+- Updated dependencies [98b2b61]
+- Updated dependencies [77f9dc4]
+- Updated dependencies [77f9dc4]
+    - @o2s/ui@0.12.0
+    - @o2s/api-harmonization@0.18.0
+
+## 0.17.0
+
+### Minor Changes
+
+- 35eeac7: implement service details page
+- 92be116: added Price model, services page implemented
+- 92be116: implement services page
+- 35eeac7: implement service details page
+- 52b3e0a: add tooltips to mocked buttons
+
+### Patch Changes
+
+- Updated dependencies [35eeac7]
+- Updated dependencies [92be116]
+- Updated dependencies [92be116]
+- Updated dependencies [35eeac7]
+- Updated dependencies [52b3e0a]
+    - @o2s/api-harmonization@0.17.0
+    - @o2s/ui@0.11.0
+
 ## 0.16.0
 
 ### Minor Changes